moved trim decimals functions to the format utils file
